requirements.txt
transformers==4.57.3
torch==2.9.1
torchvision==0.24.1
requests==2.32.5
beautifulsoup4==4.14.3
openrouter==0.1.1
datasets==4.4.1
accelerate==1.12.0
peft==0.18.0
llama-cpp-python==0.3.16
Pillow==12.0.0
opencv-python-headless==4.12.0.88
pytube==15.0.0
pytest
git+https://github.com/openai/CLIP.git
